Some small changes
1. They keydown events fire at some frequency while the key is held.
Sometimes this actually causes the rover to try to parse two base
station messages at once, which due to a bug in our parsing code (I
guess) causes an error. I added some checks so that we only fire new
drive commands when the status of the `keys` map actually changes.

2. The left_right command seemed to have a backwards sign. (Pressing 'A'
should turn the rover counterclockwise.)